# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Food Delivery service API
  description: |-
    Toptal test project Food Delivery service API docs.
  version: '1.0'
servers:
  - url: http://localhost:3000
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderItem:
      type: object
      properties:
        meal_uuid:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
        price:
          type: number
        title:
          type: string
    CreateOrderItem:
      type: object
      properties:
        meal_uuid:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
    CreateOrder:
      type: object
      properties:
        order_items:
          type: array
          description: The order items
          items: 
            $ref: '#/components/schemas/CreateOrderItem'
        coupon_code:
          type: string
          description: Coupon Code
        tip_amount:
          type: integer
          description: 'Tip amount'
          minimum: 0
    Order:
      type: object
      properties:
        order_items:
          type: array
          description: The order items
          items:
            $ref: '#/components/schemas/OrderItem'
          example:
            - name: "Pizza"
              quantity: 2
              price: 10.99
            - name: "Drink"
              quantity: 1
              price: 2.50
        coupon_code:
          type: string
          description: Coupon Code
          example: "SPECIAL20"
        tip_amount:
          type: integer
          description: 'Tip amount'
          minimum: 0
          example: 5
        restaurant_uuid:
          type: string
          format: uuid
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
...
  responses:
    BadRequest:
      description: 'Bad Request'
    Forbidden:
      description: 'Forbidden'
    NotFound:
      description: 'Not Found'
paths:
...
  /orders:
    post:
      tags: [Orders]
      summary: Create Order
      description: Creates a new order with validated details. For Admin or customer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrder'
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Orders]
      summary: Get orders
      description: Fetches list of orders. For Admin or customer.
      parameters:
        - name: page
          in: query
          required: false
          description: Zero-indexed page number. Defaults to 0
          schema:
            type: integer
            minimum: 0
            example: 0
        - name: limit
          in: query
          required: false
          description: Number of orders to retrieve per page. Defaults to 20.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            example: 20
        - name: customer_uuid
          in: query
          required: false
          description: UUID of the customer (required for customer-specific orders)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    uuid:
                      type: string
                    customer_uuid:
                      type: string
                    restaurant_uuid:
                      type: string
                    status:
                      type: string
                    total_price:
                      type: number
                    tip_amount:
                      type: number
                    discount_percentage:
                      type: integer
                    coupon_code:
                      type: string
                    created_at:
                      type: string
                      format: date-time
              example:
                - uuid: "b619ebe4-9385-4737-aeca-e6b64347fb81"
                  customer_uuid: "6a1f8646-eb13-47b7-9021-08d94fc5190c"
                  restaurant_uuid: "795d84c5-81cb-4e6a-b5d0-88291fb1e6c9"
                  status: "placed"
                  total_price: 0
                  tip_amount: 0
                  discount_percentage: 100
                  coupon_code: "SPECIAL10"
                  created_at: "2024-04-29T17:39:40.848Z"
                - uuid: "a674344c-87a5-49da-ba77-baf5a5260514"
                  customer_uuid: "6a1f8646-eb13-47b7-9021-08d94fc5190c"
                  restaurant_uuid: "795d84c5-81cb-4e6a-b5d0-88291fb1e6c9"
                  status: "placed"
                  total_price: 0.9
                  tip_amount: 0
                  discount_percentage: 10
                  coupon_code: "SPECIAL10"
                  created_at: "2024-04-29T17:38:39.194Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /orders/{uuid}:
    get:
      tags:
        - Orders
      summary: Get details of a single order
      description: Get details of a specific order by UUID.
      parameters:
        - in: path
          name: uuid
          required: true
          description: UUID of the order
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Orders]
      summary: Update order status
      description: Update status of a specific order by UUID.
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the order to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  description: Status to be updated
                  example: processing
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
              example:
                uuid: "order123"
                customer_uuid: "customer123"
                restaurant_uuid: "abc123"
                status: "completed"
                total_price: 50.00
                tip_amount: 5.00
                discount_percentage: 10
                coupon_code: "SAVE10"
                created_at: "2024-04-23T12:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
  /orders/{uuid}/status:
    get:
      tags: [Orders]
      summary: Get order status
      description: Get status of a specific order by UUID.
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the order
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  updated_at:
                    type: string
                    format: date-time
              example:
                status: "completed"
                updated_at: "2024-04-30T12:00:00Z"
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
  /orders/{uuid}/history:
    get:
      tags: [Orders]
      summary: Get order history
      description: Get history of changes for a specific order by UUID.
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the order to get history for
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    status: 
                      type: string
                    changed_at:
                      type: string
                      format: date-time
              example:
                - status: "completed"
                  changed_at: "2024-04-23T12:00:00Z"
                - status: "cancelled"
                  changed_at: "2024-04-24T12:00:00Z"
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
...